+++ src/components/commandrm/contacts/ContactBackgroundInfo.tsx
@@ -13,7 +13,7 @@ import type { Contact } from "../types";
 export const ContactBackgroundInfo = () => {
   const record = useRecordContext<Contact>();
   const translate = useTranslate();
-  const [locale = "en"] = useLocaleState();
+  const [locale = "pt-BR"] = useLocaleState();
   const { identity } = useGetIdentity();
   const isCurrentUser = record?.sales_id === identity?.id;
   const salesName = useGetSalesName(record?.sales_id, {